DataController removal changes (#996)
part of vaadin/flow#8344
