+++ vaadin-grid-flow/src/main/java/com/vaadin/flow/component/grid/Grid.java
@@ -2339,7 +2339,7 @@ public class Grid<T> extends Component
     public GridListDataView<T> getListDataView() {
         if (getDataProvider() instanceof ListDataProvider) {
             if (dataView == null || !(dataView instanceof ListDataView)) {
-                dataView = new GridListDataView<>(this);
+                dataView = new GridListDataView<>(dataCommunicator, this);
             }
             return (GridListDataView) dataView;
         }
